requirements.txt
Flask>=2.0.0
torch>=2.0.0
torchvision>=0.15.0
Pillow>=9.0.0
numpy>=1.20.0
opencv-python>=4.5.0
Werkzeug>=2.0.0
